Make the example code more structured.
